Fix the order of operations and variable name in the average function
The average function was not calculating the average correctly due to a
missing pair of parentheses around the sum of the three numbers, causing
the division to be done first. Additionally, a variable name was mistyped
in the main function, which has been fixed.

This commit also reorders the modifiers in the output statement to set
the precision before printing the average.

Signed-off-by: Tyler Ullery